server/requirements.txt
eventlet
Flask
Flask-SocketIO
Flask-MySQLDB
PyYAML